requirements.txt
streamlit>=1.33,<2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pypdf>=4.2.0
python-dotenv>=1.0.1
scikit-learn>=1.5.0
transformers>=4.44.0
torch>=2.2.0
